Extract VersionedActiveTagMatcher to common and use it with createrepo_c
